better beer user model with some helper functions
