--- beeruser.py
@@ -1,18 +1,0 @@
-import datetime
-import logging
-import os
-import webapp2
-
-from google.appengine.ext.webapp import template
-from google.appengine.api import users
-from google.appengine.ext import ndb
-from google.appengine.api import images
-from google.appengine.ext import blobstore
-from google.appengine.ext.webapp import blobstore_handlers
-
-###############################################################################
-class BeerUser(ndb.Model):
-  address = ndb.StringProperty()
-  balance = ndb.FloatProperty()
-
-
